Refactored project structure to modular
